refactor(vertex-template): adapter contract v1, delete core/apis
The adapter is now the single integration point between the UI and any
data source; the AirQo HTTP client layer is gone from the template.

- Add core/adapters/contracts/: hand-written capability interfaces
  (devices, sites, cohorts, manufacturers, users, organizations,
  access control — 69 methods) plus their request/response types,
  composed into VertexAdapter with no typeof core/apis imports
- Rename the network entity to Manufacturer (getManufacturers,
  ManufacturerCreationRequest, ...) with deprecated Network aliases;
  wire-format field names (net_*) are unchanged and their
  neutralization is documented as the v2 contract task
- Rewire all consumers (hooks, auth provider, NextAuth options,
  components, API routes) to the adapter and contract types
- Move feedback and cloudinary from core/apis to core/services —
  they are app services, not data-adapter concerns
- Delete core/apis/* including the axios client config

Verified: tsc, eslint (0 errors on diff and contracts), next build,
and a browser walkthrough of home, my-devices, admin manufacturers,
and cohorts on mock data with zero console errors.
