--- src/vertex-template/core/apis/devices.ts
@@ -1,613 +1,0 @@
-import createSecureApiClient from "../utils/secureApiProxyClient";
-import type {
-  DevicesSummaryResponse,
-  DeviceAvailabilityResponse,
-  DeviceClaimRequest,
-  DeviceClaimResponse,
-  BulkDeviceClaimRequest,
-  BulkDeviceClaimResponse,
-  MyDevicesResponse,
-  DeviceAssignmentRequest,
-  DeviceAssignmentResponse,
-  Device,
-  DeviceCreationResponse,
-  BulkImportDeviceResponse,
-  DeviceUpdateGroupResponse,
-  MaintenanceLogData,
-  DecryptionRequest,
-  DecryptionResponse,
-} from "@/app/types/devices";
-
-const jwtApiClient = createSecureApiClient();
-
-interface DeviceStatusSummary {
-  _id: string;
-  created_at: string;
-  total_active_device_count: number;
-  count_of_online_devices: number;
-  count_of_offline_devices: number;
-  count_of_mains: number;
-  count_of_solar_devices: number;
-  count_of_alternator_devices: number;
-  count_due_maintenance: number;
-  count_overdue_maintenance: number;
-  count_unspecified_maintenance: number;
-  online_devices: Device[];
-  offline_devices: Device[];
-}
-
-export interface DeviceStatusResponse {
-  message: string;
-  data: DeviceStatusSummary[];
-}
-
-export interface DeviceCountSummary {
-  total_monitors: number;
-  operational: number;
-  transmitting: number;
-  not_transmitting: number;
-  data_available: number;
-}
-
-export interface DeviceCountResponse {
-  success: boolean;
-  message: string;
-  summary: DeviceCountSummary;
-}
-
-export interface DeviceDetailsResponse {
-  message: string;
-  data: Device;
-}
-
-export interface DeviceActivity {
-  _id: string;
-  device: string;
-  date: string;
-  description: string;
-  activityType: string;
-  activity_by?: {
-    user_id: string;
-    name: string;
-    email: string;
-  };
-  site_id?: string;
-  network?: string;
-  nextMaintenance?: string;
-  deployment_type?: string;
-  tags?: string[];
-  createdAt: string;
-  updatedAt?: string;
-}
-
-export interface DeviceActivitiesResponse {
-  success: boolean;
-  message: string;
-  site_activities: DeviceActivity[];
-  meta: {
-    total: number;
-    limit: number;
-    skip: number;
-    page: number;
-    totalPages: number;
-  };
-}
-
-export type GetDevicesSummaryParams = Partial<Device> & {
-  limit?: number;
-  skip?: number;
-  search?: string;
-  sortBy?: string;
-  order?: "asc" | "desc";
-  group?: string;
-};
-
-export const devices = {
-  getDevicesSummaryApi: async (params: GetDevicesSummaryParams, signal?: AbortSignal) => {
-    try {
-      const queryParams = new URLSearchParams();
-
-      Object.entries(params).forEach(([key, value]) => {
-        if (value !== undefined && value !== null) {
-          queryParams.set(key, String(value));
-        }
-      });
-
-      const response = await jwtApiClient.get<DevicesSummaryResponse>(
-        `/devices/summary?${queryParams.toString()}`,
-        { headers: { "X-Auth-Type": "JWT" }, signal }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getDevicesByCohorts: async (params: Partial<Device> & {
-    cohort_ids: string[];
-    limit?: number;
-    skip?: number;
-    search?: string;
-    sortBy?: string;
-    order?: "asc" | "desc";
-  }, signal?: AbortSignal) => {
-    try {
-      const { cohort_ids, ...rest } = params;
-      const queryParams = new URLSearchParams();
-      Object.entries(rest).forEach(([key, value]) => {
-        if (value !== undefined && value !== null) {
-          queryParams.set(key, String(value));
-        }
-      });
-
-      const response = await jwtApiClient.post<DevicesSummaryResponse>(
-        `/devices/cohorts/cached-devices?${queryParams.toString()}`,
-        { cohort_ids },
-        { headers: { "X-Auth-Type": "JWT" }, signal }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getDevicesByStatusApi: async (params: {
-    status: string;
-    cohort_id?: string[];
-    limit?: number;
-    skip?: number;
-    search?: string;
-    sortBy?: string;
-    order?: "asc" | "desc";
-    network?: string;
-  }) => {
-    try {
-      const { status, cohort_id, ...rest } = params;
-      const queryParams = new URLSearchParams();
-
-      if (cohort_id && cohort_id.length > 0) {
-        queryParams.set("cohort_id", cohort_id.join(","));
-      }
-
-      Object.entries(rest).forEach(([key, value]) => {
-        if (value !== undefined && value !== null) {
-          queryParams.set(key, String(value));
-        }
-      });
-
-      const formattedStatus = status.replace(/_/g, '-').replace(/ /g, '-');
-      const response = await jwtApiClient.get<DevicesSummaryResponse>(
-        `/devices/status/${formattedStatus}?${queryParams.toString()}`,
-        { headers: { "X-Auth-Type": "JWT" } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-  getDeviceCountApi: async (params: {
-    cohort_id?: string[];
-    network?: string;
-  }): Promise<DeviceCountResponse> => {
-    try {
-      const { cohort_id, network } = params;
-      const queryParams = new URLSearchParams();
-
-      if (cohort_id && cohort_id.length > 0) {
-        queryParams.set("cohort_id", cohort_id.join(","));
-      }
-
-      if (network) {
-        queryParams.set("network", network);
-      }
-
-      const response = await jwtApiClient.get<DeviceCountResponse>(
-        `/devices/summary/count?${queryParams.toString()}`,
-        { headers: { "X-Auth-Type": "JWT" } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  checkDeviceAvailability: async (deviceName: string): Promise<DeviceAvailabilityResponse> => {
-    try {
-      const response = await jwtApiClient.get<DeviceAvailabilityResponse>(
-        `/devices/check-availability/${deviceName}`,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  claimDevice: async (claimData: DeviceClaimRequest): Promise<DeviceClaimResponse> => {
-    try {
-      const response = await jwtApiClient.post<DeviceClaimResponse>(
-        `/devices/claim`,
-        claimData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  claimDevicesBulk: async (claimData: BulkDeviceClaimRequest): Promise<BulkDeviceClaimResponse> => {
-    try {
-      const response = await jwtApiClient.post<BulkDeviceClaimResponse>(
-        `/devices/claim/bulk`,
-        claimData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getMyDevices: async (
-    userId: string,
-    groupIds?: string[],
-    cohortIds?: string[]
-  ): Promise<MyDevicesResponse> => {
-    try {
-      const params = new URLSearchParams({ user_id: userId });
-
-      if (groupIds && groupIds.length > 0) {
-        params.append("group_ids", groupIds.join(","));
-      }
-
-      if (cohortIds && cohortIds.length > 0) {
-        params.append("cohort_ids", cohortIds.join(","));
-      }
-
-      const response = await jwtApiClient.get<MyDevicesResponse>(
-        `/devices/my-devices?${params.toString()}`,
-        { headers: { "X-Auth-Type": "JWT" } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  assignDeviceToOrganization: async (assignmentData: DeviceAssignmentRequest): Promise<DeviceAssignmentResponse> => {
-    try {
-      const response = await jwtApiClient.post<DeviceAssignmentResponse>(
-        `/devices/assign-organization`,
-        assignmentData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  unassignDeviceFromOrganization: async (deviceName: string, userId: string): Promise<DeviceAssignmentResponse> => {
-    try {
-      const response = await jwtApiClient.post<DeviceAssignmentResponse>(
-        `/devices/unassign-organization`,
-        { device_name: deviceName, user_id: userId },
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getDeviceDetails: async (deviceId: string): Promise<DeviceDetailsResponse> => {
-    try {
-      const response = await jwtApiClient.get<DeviceDetailsResponse>(
-        `/devices/${deviceId}`,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getDeviceStatusFeed: async (deviceNumber: number) => {
-    try {
-      const response = await jwtApiClient.get(
-        `/devices/feeds/transform/recent?channel=${deviceNumber}`,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  deployDevice: async (deviceData: {
-    deviceName: string;
-    height: string;
-    mountType: string;
-    powerType: string;
-    isPrimaryInLocation: boolean;
-    latitude: string;
-    longitude: string;
-    site_name?: string;
-    site_id?: string;
-    network: string;
-    user_id: string;
-    deployment_date: string | undefined;
-    firstName?: string;
-    lastName?: string;
-    email?: string;
-    userName?: string;
-  }) => {
-    try {
-      const toIso = (d?: string) =>
-        d ? new Date(d).toISOString() : new Date().toISOString();
-      const latitude = Number(deviceData.latitude);
-      const longitude = Number(deviceData.longitude);
-      const height = Number(deviceData.height);
-      if (Number.isNaN(latitude) || Number.isNaN(longitude) || Number.isNaN(height)) {
-        throw new Error("Invalid numeric values for latitude, longitude or height.");
-      }
-      const deploymentPayload = [{
-        date: toIso(deviceData.deployment_date),
-        mountType: deviceData.mountType,
-        powerType: deviceData.powerType,
-        isPrimaryInLocation: deviceData.isPrimaryInLocation,
-        latitude,
-        longitude,
-        ...(deviceData.site_id
-          ? { site_id: deviceData.site_id, site_name: deviceData.site_name || `${deviceData.deviceName} Site` }
-          : { site_name: deviceData.site_name || `${deviceData.deviceName} Site` }),
-        network: deviceData.network,
-        deviceName: deviceData.deviceName,
-        height,
-        user_id: deviceData.user_id,
-        firstName: deviceData.firstName,
-        lastName: deviceData.lastName,
-        email: deviceData.email,
-        userName: deviceData.userName
-      }];
-
-      const response = await jwtApiClient.post(
-        `/devices/activities/deploy/batch`,
-        deploymentPayload,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  recallDevice: async (deviceName: string, recallData: {
-    recallType: string;
-    user_id: string;
-    date: string;
-    firstName?: string;
-    lastName?: string;
-    email?: string;
-    userName?: string;
-  }) => {
-    try {
-      const response = await jwtApiClient.post(
-        `/devices/activities/recall?deviceName=${deviceName}`,
-        recallData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  createDevice: async (deviceData: {
-    long_name: string;
-    category: string;
-    description?: string;
-    network: string;
-    tags?: string[];
-  }): Promise<DeviceCreationResponse> => {
-    try {
-      const response = await jwtApiClient.post(
-        `/devices`,
-        deviceData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  importDevice: async (deviceData: {
-    long_name: string;
-    category: string;
-    network: string;
-    device_number?: string;
-    writeKey?: string;
-    readKey?: string;
-    description?: string;
-    serial_number: string;
-    api_code?: string;
-    cohort_id?: string;
-    user_id: string;
-    tags?: string[];
-  }): Promise<DeviceCreationResponse> => {
-    try {
-      const response = await jwtApiClient.post(
-        `/devices/soft`,
-        deviceData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  importBulkDevicesCSV: async (
-    formData: FormData
-  ): Promise<BulkImportDeviceResponse> => {
-    try {
-      const response = await jwtApiClient.post(
-  `/devices/soft/bulk`,
-  formData,
-  {
-    headers: {
-      'X-Auth-Type': 'JWT',
-    },
-  }
-);
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  importBulkDevicesJSON: async (
-    payload: {
-      user_id: string;
-      cohort_id?: string;
-      network_override?: string;
-      devices: Array<{
-        long_name: string;
-        serial_number: string;
-        network?: string;
-        latitude?: number;
-        longitude?: number;
-        api_code?: string;
-        category?: string;
-        description?: string;
-        device_number?: number;
-        tags?: string[];
-      }>;
-    }
-  ): Promise<BulkImportDeviceResponse> => {
-    try {
-      const response = await jwtApiClient.post(
-        `/devices/soft/bulk`,
-        payload,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  updateDeviceLocal: async (deviceId: string, deviceData: Partial<Device>) => {
-    try {
-      const { ...updateData } = deviceData;
-      const response = await jwtApiClient.put(
-        `/devices?id=${deviceId}`,
-        updateData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  updateDeviceGlobal: async (deviceId: string, deviceData: Partial<Device>) => {
-    try {
-      const { ...updateData } = deviceData;
-      const response = await jwtApiClient.put(
-        `/devices/soft?id=${deviceId}`,
-        updateData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  addMaintenanceLog: async (deviceName: string, logData: MaintenanceLogData) => {
-    try {
-      const response = await jwtApiClient.post(
-        `/devices/activities/maintain?deviceName=${deviceName}`,
-        logData,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  bulkUpdateDeviceDetails: async (
-    deviceIds: string[],
-    updateData: Record<string, unknown>
-  ): Promise<DeviceUpdateGroupResponse> => {
-    try {
-      const sanitizedDeviceIds = deviceIds.map((id) => id.trim()).filter(Boolean);
-      if (sanitizedDeviceIds.length === 0) {
-        throw new Error("At least one device ID is required.");
-      }
-
-      const requestBody = {
-        deviceIds: sanitizedDeviceIds,
-        updateData,
-      };
-
-      const response = await jwtApiClient.put(`/devices/bulk`, requestBody, {
-        headers: { "X-Auth-Type": "JWT" },
-      });
-
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  decryptDeviceKeys: async (
-    requestBody: DecryptionRequest[]
-  ): Promise<DecryptionResponse> => {
-    try {
-      const response = await jwtApiClient.post<DecryptionResponse>(
-        `/devices/decrypt/bulk`,
-        requestBody,
-        { headers: { "X-Auth-Type": "JWT" } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getOrphanedDevices: async (userId: string) => {
-    try {
-      const params = new URLSearchParams({ user_id: userId });
-      const response = await jwtApiClient.get(
-        `/devices/orphaned?${params.toString()}`,
-        { headers: { "X-Auth-Type": "JWT" } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-
-  getDeviceActivities: async (deviceName: string, params: { page?: number; limit?: number } = {}): Promise<DeviceActivitiesResponse> => {
-    try {
-      const { page = 1, limit = 10 } = params;
-      const queryParams = new URLSearchParams({
-        device: deviceName,
-        page: String(page),
-        limit: String(limit)
-      });
-      const response = await jwtApiClient.get<DeviceActivitiesResponse>(
-        `/devices/activities?${queryParams.toString()}`,
-        { headers: { 'X-Auth-Type': 'JWT' } }
-      );
-      return response.data;
-    } catch (error) {
-      throw error;
-    }
-  },
-};
